Phase 1: Rebrand Riecoin to Freycoin
- Rename all binaries: riecoind -> freycoind, riecoin-qt -> freycoin-qt, etc.
- Update CMakeLists.txt project name and targets
- Rename library targets: riecoin_* -> freycoin_*
- Update Qt target names: riecoinqt -> freycoinqt
- Update bech32 HRP: ric -> frey, tric -> tfrey, rric -> rfrey
- Rename source files to match new naming convention
- Update copyright headers
- Update documentation and config files
- Add Qt D3D patches for MinGW cross-compilation
- Fix GMP cross-compilation with CC_FOR_BUILD

+++ src/test/skiplist_tests.cpp
@@ -1,5 +1,5 @@
 // Copyright (c) 2014-2022 The Bitcoin Core developers
-// Copyright (c) 2013-present The Riecoin developers
+// Copyright (c) 2013-present The Freycoin developers
 // Distributed under the MIT software license, see the accompanying
 // file COPYING or http://www.opensource.org/licenses/mit-license.php.
 
@@ -118,13 +118,13 @@ BOOST_AUTO_TEST_CASE(findearliestatleast_test)
             int64_t medianTimePast = vBlocksMain[i].GetMedianTimePast();
             int r{int(m_rng.randrange(medianTimePast))};
             vBlocksMain[i].nTime = uint32_t(r + medianTimePast);
-            vBlocksMain[i].nTimeMax = std::max(vBlocksMain[i].nTime, vBlocksMain[i-1].nTimeMax);
+            vBlocksMain[i].nTimeMax = std::max(static_cast<int64_t>(vBlocksMain[i].nTime), vBlocksMain[i-1].nTimeMax);
         }
     }
     // Check that we set nTimeMax up correctly.
     int64_t curTimeMax(0);
     for (unsigned int i=0; i<vBlocksMain.size(); ++i) {
-        curTimeMax = std::max(curTimeMax, vBlocksMain[i].nTime);
+        curTimeMax = std::max(curTimeMax, static_cast<int64_t>(vBlocksMain[i].nTime));
         BOOST_CHECK(curTimeMax == vBlocksMain[i].nTimeMax);
     }
